abstract access check methods to independent class
