+++ src/main/java/com/revlearn/team1/service/accessControl/AccessControlService.java
@@ -1,0 +1,9 @@
+package com.revlearn.team1.service.accessControl;
+
+import com.revlearn.team1.model.Course;
+
+public interface AccessControlService {
+    boolean verifyStudentLevelAccess(Course course);
+    boolean verifyEducatorLevelAccess(Course course);
+
+    }
